requirements.txt
# requirements.txt

annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
-e git+https://github.com/matteobettini/BenchMARL.git@a115079b9f8d13f2ebf77ec49065b0f117e8df27#egg=benchmarl
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
docutils==0.21.2
eval_type_backport==0.2.2
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
gym==0.26.2
gym-notices==0.0.8
-e git+https://github.com/proroklab/ControllingBehavioralDiversity.git@6406ceffe2975fcd14a6b98fca229910c85b6d53#egg=het_control
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
moviepy==2.2.1
mpmath==1.3.0
# networkx==3.5
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
proglog==0.1.12
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
pyglet==1.5.27
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
sentry-sdk==2.33.2
six==1.17.0
smmap==5.0.2
# sympy==1.14.0
-e git+https://github.com/matteobettini/tensordict.git@a25b22b0e7856c6e277bd552b67b76e790a816d3#egg=tensordict
# torch==1.13.1
-e git+https://github.com/matteobettini/rl.git@a0c83d5146ef1fd405be604b047afd07996b588c#egg=torchrl
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
-e git+https://github.com/proroklab/VectorizedMultiAgentSimulator.git@4be87bb5d3eba4a6d96825b9177760c766ea8072#egg=vmas
wandb==0.21.0
zipp==3.23.0
